Fixed legato mode and removed useless param setters
Legato mode was behaving as expected : the ADSR envelope wasn't re-triggered when glide is 0. A "legato" flag was added in SineWaveVoice.
ADSR envelope now uses pointers to the AudioParameterFloat raw values. These pointers are set in the constructor so there's no need for setters anymore.
Other useless setters were removed.