Added bigger graduations for sliders
Each slider has either one bigger graduation showing the point for 0 (bend and output) or 2 bigger graduations for min and max.
A method was added in DefaultParameterValues to compute the skew factor for the output gain slider so that 0dB is centered. Also changed the default parameter values to being either min, 0 or max.