Added dynamic routing capabilities
A preprocessor directive allows to show a routing component (AudioChainComponent) for changing the order of the subsequent FXs (envelope, filter and overdrive). The audio processing code was made more flexible to allow this.
For now, changing the audio routing has no effect on the processor state (so restoring a preset won't change the routing) and the host has no access to the routing. The AudioChainComponent is the only way to change it.